Rely on property wrappers
